Refactor difficulty selection modal component and implement GET endpoint for retrieving game results
