+++ backend/controllers/memoryController.js
@@ -27,3 +27,18 @@ exports.saveGameData = async (req, res) => {
         res.status(500).json({ message: 'Error saving game data', error });
     }
 };
+
+exports.getGameHistory = async (req, res) => {
+    try {
+      const { userID } = req.query;
+  
+      const query = userID ? { userID } : {};
+  
+      const saves = await Save.find(query).sort({ gameDate: -1 });
+  
+      res.status(200).json(saves);
+    } catch (error) {
+      console.error('Error fetching game history:', error);
+      res.status(500).json({ message: 'Error fetching game history', error });
+    }
+  };
