+++ backend/routes/memoryRoutes.js
@@ -1,8 +1,10 @@
 const express = require('express');
-const { saveGameData } = require('../controllers/memoryController');
+const { saveGameData, getGameHistory } = require('../controllers/memoryController');
 const router = express.Router();
 
 // Route to save game data
 router.post('/save', saveGameData);
+
+router.get('/history', getGameHistory);
 
 module.exports = router;
